requirements.txt
ruamel.yaml==0.15.34
kubernetes==4.0.0
urllib3>=1.20
backports.ssl-match-hostname>=3.5.0.1
ipaddress>=1.0.18
oauthlib>=2.0.6
requests-oauthlib==0.8.0
git+https://github.com/magmax/python-inquirer.git
